Release 0.9.2

+++ src/backend/opengl/font_atlas_impl.hpp
@@ -71,8 +71,8 @@ struct Glyph {
     size_t mWidth;
     size_t mHeight;
 
-    size_t mBearingX;
-    size_t mBearingY;
+    int mBearingX;
+    int mBearingY;
 
     float  mAdvanceX;
     float  mAdvanceY;
